Refactor file system and lambda functions
