+++ lambda/lib/base64-handler.js
@@ -1,0 +1,12 @@
+module.exports = class Base64Handler {
+    constructor(image) {
+        let base64Regex = /data:([^/]+)\/([^;]+);base64,(.+)/;
+        let matches = base64Regex.exec(image);
+        this.mimeType = {
+            type: matches[1],
+            subtype: matches[2]
+        };
+        this.base64Image = matches[3];
+        this.buffer = Buffer.from(this.base64Image, 'base64');
+    }
+};
