+++ lambda/lib/file-writer.js
@@ -1,8 +1,11 @@
 'use strict';
 
+const AWSS3 = require('aws-sdk/clients/s3');
+const _ = require('lodash');
+
 module.exports = class FileWriter {
     constructor(s3) {
-        this.s3 = s3;
+        this.s3 = _.isUndefined(s3) ? new AWSS3() : s3;
     }
 
     saveObject(imageFile, callback) {
