+++ lambda/lib/moderation-threshold-exceeded.js
@@ -1,0 +1,6 @@
+module.exports = class ModerationThresholdExceeded extends Error {
+    constructor(message) {
+        super(message);
+        this.name = "ModerationThresholdExceeded";
+    }
+};
