+++ lambda/test/file-builder.js
@@ -2,7 +2,7 @@
 
 let expect = require('chai').expect;
 let md5 = require('md5');
-let FileBuilder = require('../file-builder');
+let FileBuilder = require('../lib/file-builder');
 let base64ImageMetadata = 'data:image/png;base64,';
 let base64Image = 'iVBORw0KGgoAAAANSUhEUgAAAAsAAAAECAYAAABY+sXzAAAABHNCSVQICAgIfAhkiAAAAFlJREFUCJl9yjEKwCAUBNERv43g/U+WKqVVsNRKxU0jJGnyYGGLcTln1VoBSCnRWsN7D0AIATOjlEKMEbQd16kxhuac6r1rrfWZJBmbAEn8eeKlT/z+zjkAbkDFRMbggmGwAAAAAElFTkSuQmCC';
 let imageMD5 = md5(Buffer.from(base64Image, 'base64'));
